Reworking security permissions...

+++ api/ops/advert.php
@@ -28,7 +28,7 @@ class ApiPage extends OpsApiPageBase
 		global $_profile;
 		
 		$club_id = (int)get_required_param('club_id');
-		$this->check_permissions($club_id);
+		check_permissions(PERMISSION_CLUB_MANAGER, $club_id);
 		$club = $_profile->clubs[$club_id];
 		
 		$raw_message = get_required_param('message');
@@ -69,7 +69,7 @@ class ApiPage extends OpsApiPageBase
 	
 	function create_op_help()
 	{
-		$help = new ApiHelp('Create advertisement.');
+		$help = new ApiHelp(PERMISSION_CLUB_MANAGER, 'Create advertisement.');
 		$help->request_param('club_id', 'Club id. Required.');
 		$help->request_param('message', 'Advertizement message text.');
 		$help->request_param('start', 'Time when the message will start apearing in the club main page. It is either unix timestamp or time string in format (php datetime format) "Y-m-d H:i" ("2018-06-23 17:33").', 'the message starts being showed immediatly.');
@@ -78,11 +78,6 @@ class ApiPage extends OpsApiPageBase
 
 		$help->response_param('advert_id', 'Newly created advertisement id.');
 		return $help;
-	}
-	
-	function create_op_permissions()
-	{
-		return PERMISSION_CLUB_MANAGER;
 	}
 
 	//-------------------------------------------------------------------------------------------------------
@@ -94,7 +89,7 @@ class ApiPage extends OpsApiPageBase
 		
 		$advert_id = (int)get_required_param('advert_id');
 		list ($club_id, $raw_message, $message, $start, $end, $lang) = Db::record(get_label('advert'), 'SELECT club_id, raw_message, message, timestamp, expires, lang FROM news WHERE id = ?', $advert_id);
-		$this->check_permissions($club_id);
+		check_permissions(PERMISSION_CLUB_MANAGER, $club_id);
 		$club = $_profile->clubs[$club_id];
 		
 		if (isset($_REQUEST['message']))
@@ -138,18 +133,13 @@ class ApiPage extends OpsApiPageBase
 	
 	function change_op_help()
 	{
-		$help = new ApiHelp('Change advertisement.');
+		$help = new ApiHelp(PERMISSION_CLUB_MANAGER, 'Change advertisement.');
 		$help->request_param('advert_id', 'Advertisement id.');
 		$help->request_param('message', 'Advertizement message text.', 'remains the same.');
 		$help->request_param('start', 'Time when the message will start apearing in the club main page. It is either unix timestamp or time string in format (php style) "Y-m-d H:i" ("2018-06-23 17:33").', 'the message starts being showed immediatly.', 'remains the same.');
 		$help->request_param('end', 'Expiration time.  It is either unix timestamp or time string in format (php style) "Y-m-d H:i" ("2018-06-23 17:33").', 'remains the same.');
 		$help->request_param('lang', 'Advertisement language. 1 (English) or 2 (Russian). Other languages are not supported yet.<br><dfn>When 0:</dfn> auto-detected by analyzing message character codes.', 'remains the same.');
 		return $help;
-	}
-	
-	function change_op_permissions()
-	{
-		return PERMISSION_CLUB_MANAGER;
 	}
 
 	//-------------------------------------------------------------------------------------------------------
@@ -159,7 +149,7 @@ class ApiPage extends OpsApiPageBase
 	{
 		$advert_id = (int)get_required_param('advert_id');
 		list ($club_id) = Db::record(get_label('advert'), 'SELECT club_id FROM news WHERE id = ?', $advert_id);
-		$this->check_permissions($club_id);
+		check_permissions(PERMISSION_CLUB_MANAGER, $club_id);
 		
 		Db::begin();
 		Db::exec(get_label('advert'), 'DELETE FROM news WHERE id = ?', $advert_id);
@@ -169,14 +159,9 @@ class ApiPage extends OpsApiPageBase
 	
 	function delete_op_help()
 	{
-		$help = new ApiHelp('Delete advertisement.');
+		$help = new ApiHelp(PERMISSION_CLUB_MANAGER, 'Delete advertisement.');
 		$help->request_param('advert_id', 'Advertisement id.');
 		return $help;
-	}
-	
-	function delete_op_permissions()
-	{
-		return PERMISSION_CLUB_MANAGER;
 	}
 }
